requirements.txt
Pillow>=9.5.0
opencv-python>=4.8.1
numpy>=1.25.0
pandas>=2.1.0
openpyxl>=3.1.0
python-dateutil>=2.8.2
